move all options to it own files
